More find() tests into separate file
